Master

+++ test/compilationTests/gnosis/Utils/Math.sol
@@ -94,12 +94,12 @@ library Math {
         zpow = zpow * z / ONE;
         result += 0x9c7 * zpow / ONE;
         if (shift >= 0) {
-            if (result >> (256-shift) > 0)
-                return (2**256-1);
-            return result << shift;
+            if (result >> uint(256 - shift) > 0)
+                return (2 ** 256 - 1);
+            return result << uint(shift);
         }
         else
-            return result >> (-shift);
+            return result >> uint(-shift);
     }
 
     /// @dev Returns natural logarithm value of given x
@@ -154,7 +154,7 @@ library Math {
 
     /// @dev Returns base 2 logarithm value of given x
     /// @param x x
-    /// @return logarithmic value
+    /// @return lo logarithmic value
     function floorLog2(uint x)
         public
         pure
@@ -175,7 +175,7 @@ library Math {
 
     /// @dev Returns maximum of an array
     /// @param nums Numbers to look through
-    /// @return Maximum number
+    /// @return max Maximum number
     function max(int[] memory nums)
         public
         pure
